make fake api async
