+++ src/layouts/utils/helpers.js
@@ -27,25 +27,3 @@ export const addCards = (num, cardType, projectCardType, titleVisibility) => {
     ))
   }
 }
-
-export function fakeCallToAPI(length = 7, data = ['any', 'any', 'any', 'any'], title='') {
-  // data: type, role, level, location
-  const response = Array.from({ length: length }, () => {
-    return {
-      title: title? title : faker.commerce.productName(),
-      type: data[0],
-      role: data[1],
-      level: data[2],
-      location: data[3],
-      description: faker.hacker.phrase()
-    }
-  })
-  return response
-}
-
-// fake API for search
-
-export const fakeAPISearch = str => {
-  const random = Math.floor(Math.random()*11)
-  return fakeCallToAPI(random, undefined, str)
-}
